requirements.txt
annotated-types==0.6.0
anyio==3.7.1
bcrypt==4.1.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
databases==0.8.0
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.1.0.post1
fastapi==0.106.0
greenlet==3.0.3
h11==0.14.0
idna==3.6
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql-connector-python==8.2.0
passlib==1.7.4
Pillow==10.1.0
protobuf==4.21.12
pyasn1==0.5.1
pydantic==2.5.3
pydantic_core==2.14.6
PyMySQL==1.1.0
pypng==0.20220715.0
python-jose==3.3.0
python-multipart==0.0.6
pyzbar==0.1.9
qrcode==7.4.2
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.50
starlette==0.27.0
typing_extensions==4.9.0
urllib3==2.1.0
uvicorn==0.25.0
validators==0.22.0
